server/requirements.txt
beautifulsoup4
psycopg
python-dotenv
psycopg-pool
fastapi[standard]
pydantic